requirements.txt
fastapi==0.118.0
uvicorn==0.37.0
openai==1.109.1
azure-cosmos==4.9.0
httpx==0.28.1
python-dotenv==1.1.1
serpapi>=0.1.5
